requirements.txt
allure-pytest==2.15.2
allure-python-commons==2.15.2
attrs==25.4.0
greenlet==3.3.0
iniconfig==2.3.0
packaging==25.0
playwright==1.57.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
typing_extensions==4.15.0
